Move zendesk service to clients
