+++ libs/api/domains/communications/src/lib/communications.service.spec.ts
@@ -1,7 +1,7 @@
 import { EmailModule, EmailService } from '@island.is/email-service'
 import { LoggingModule } from '@island.is/logging'
 import { Test } from '@nestjs/testing'
-import { ZendeskModule, ZendeskService } from '@island.is/zendesk-service'
+import { ZendeskModule, ZendeskService } from '@island.is/clients/zendesk'
 import { CommunicationsService } from './communications.service'
 import { ContactUsInput } from './dto/contactUs.input'
 import { TellUsAStoryInput } from './dto/tellUsAStory.input'
