+++ libs/api/domains/communications/src/lib/communications.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable } from '@nestjs/common'
 import { EmailService } from '@island.is/email-service'
-import { ZendeskService } from '@island.is/zendesk-service'
+import { ZendeskService } from '@island.is/clients/zendesk'
 import { Logger, LOGGER_PROVIDER } from '@island.is/logging'
 import { SendMailOptions } from 'nodemailer'
 import { ContactUsInput } from './dto/contactUs.input'
