+++ workspace.json
@@ -2319,6 +2319,27 @@
         }
       }
     },
+    "clients-zendesk": {
+      "root": "libs/clients/zendesk",
+      "sourceRoot": "libs/clients/zendesk/src",
+      "projectType": "library",
+      "architect": {
+        "lint": {
+          "builder": "@nrwl/linter:eslint",
+          "options": {
+            "lintFilePatterns": ["libs/clients/zendesk/**/*.ts"]
+          }
+        },
+        "test": {
+          "builder": "@nrwl/jest:jest",
+          "outputs": ["coverage/libs/clients/zendesk"],
+          "options": {
+            "jestConfig": "libs/clients/zendesk/jest.config.js",
+            "passWithNoTests": true
+          }
+        }
+      }
+    },
     "content-search-index-manager": {
       "root": "libs/content-search-index-manager",
       "sourceRoot": "libs/content-search-index-manager/src",
@@ -6260,27 +6281,6 @@
             "linter": "eslint",
             "tsConfig": ["apps/web-e2e/tsconfig.e2e.json"],
             "exclude": ["**/node_modules/**", "!apps/web-e2e/**/*"]
-          }
-        }
-      }
-    },
-    "zendesk-service": {
-      "root": "libs/zendesk-service",
-      "sourceRoot": "libs/zendesk-service/src",
-      "projectType": "library",
-      "architect": {
-        "lint": {
-          "builder": "@nrwl/linter:eslint",
-          "options": {
-            "lintFilePatterns": ["libs/zendesk-service/**/*.ts"]
-          }
-        },
-        "test": {
-          "builder": "@nrwl/jest:jest",
-          "outputs": ["coverage/libs/zendesk-service"],
-          "options": {
-            "jestConfig": "libs/zendesk-service/jest.config.js",
-            "passWithNoTests": true
           }
         }
       }
